pyproject.toml
[project]
name = "secret-santa-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cognee>=0.4.1",
    "fastapi>=0.124.1",
    "google-generativeai>=0.8.5",
    "gradio>=5.23.1",
    "networkx>=3.6.1",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "plotly>=6.5.0",
    "protego>=0.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyvis>=0.3.2",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.8",
]
